Fix another sign comparison case ...

+++ cpp/util/thread_pool.cc
@@ -142,7 +142,7 @@ ThreadPool::ThreadPool()
 
 
 ThreadPool::ThreadPool(size_t num_threads) : impl_(new Impl) {
-  CHECK_GT(num_threads, 0);
+  CHECK_GT(num_threads, static_cast<size_t>(0));
   LOG(INFO) << "ThreadPool starting with " << num_threads << " threads";
   for (int i = 0; i < static_cast<int64_t>(num_threads); ++i)
     impl_->threads_.emplace_back(thread(&Impl::Worker, impl_.get()));
